backend/pyproject.toml
[project]
name = "sme-score-api"
version = "1.0.0"
description = "API de scoring et certification des PME ivoiriennes"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "qrcode[pil]>=7.4.2",
    "pandas>=2.2.0",
    "numpy>=1.26.4",
    "scikit-learn>=1.4.1",
    "joblib>=1.3.2",
    "faker>=22.6.0",
    "supabase>=2.3.4",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.1",
    "gunicorn>=21.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
